move the `SourceSinkConverter` tests to a separate test file
